Extract file rename workflow
