+++ src/gui_app/folder_browser.rs
@@ -182,6 +182,8 @@ mod file_selection;
 
 mod file_view_window;
 
+mod file_rename_workflow;
+
 mod file_columns;
 #[cfg(test)]
 pub(super) use file_columns::MIN_FILE_COLUMN_WIDTH;
